Add files via upload
It is a College Library Portal only accessible by college administration developed to facilitate the library book issue system to manage the information of book that has been issued, misplaced, not returned by the student, and available to issue. It also manages fine levied on the student who has not returned the book on time.

It aims in providing an easy, quick way to manage the college library book system. This portal will allow the administrator to see the information of books issued, misplaced, and fine on the student. The portal provides an option to keep a record of which student issued which book with the help of a book id. The administrator can log in and use the portal to store student info on who has issued the book and to keep a record of who has not returned the book.

Admin will be responsible for maintaining the portal and will also have access to the whole system.

[By using html, css, javascript, database , php]